# src/main/java/ru/skypro/homework/dto/FullAdsDto.java
package ru.skypro.homework.dto;

import lombok.*;

@Data
public class FullAdsDto {

    private String image;
    private String authorLastName;
    private String authorFirstName;
    private String phone;
    private int price;
    private String description;
    private int pk;
    private String title;
    private String email;
}